+++ src/ua/src/ua.js
@@ -5,5 +5,5 @@ KISSY.add("ua", function (S, UA) {
     S.UA = UA;
     return UA;
 }, {
-    requires:["ua/extra"]
+    requires:["ua/extra", "ua/css"]
 });